track skills on umas as a map from groupId to skillId, rather than just skillIds
this simplifies some code and improves a lot of things. specifically, it now removes skills of the same group on the compare uma as the skill under test (i.e. if the base/standard has a white and we're testing the gold version, it now replaces the white with the gold rather than adding the gold on top of it).

as a side effect this greatly cleans up how allowing multiple debuffs is handled (instead of faking skill ids we can fake group ids, which is much simpler). because of that we can get rid of all the skillmeta/skilldata wrappers that split on -. this does also entail getting rid of split alternatives for seirios in the skilltools repo, which has been done concurrently (as that was always a bit of a headache and is no longer really necessary now that the umalator replaced the old spreadsheets (though the umalator does not yet allow selecting popularity)).

also as a consequence this removes immutable-sorted, which reduces the bundle size a fair bit.