revert 998fafdd33f829e9f44b9856388d97dceade5b1d
i realized after implementing it that i dont actually like it, but wanted to save the work. the problem is that it opens a real can of worms. currently it only runs when double-clicking from the chart and not adding skills manually; which is sort of weird and sloppy. that can be fixed by diffing any state changes of uma1 and rerunning the chart, but the real issue is that adding skills (particularly accel skills) affects the values of every other skill and not just ones in the group. we want to rerun the group because their cost changes (and therefore sp efficiency) but the sp efficiency of everything else changes also. we _could_ run every skill in the background with a small sample count, look at what actually changes, and then do a full rerun of the ones that do, but then you have the issue of the graph will show the new skill on it for ones that changed but not for ones that don't. overall i think there's not a great solution to this and the full table should be rerun when a skill is added or removed. i think instead what i'm going to do is 1. not update the cost (and hence sp efficiency) when a skill is added 2. not hide the skill that was added (so that both it and other skills in the group are still present, which makes the cost non-update look better i think) 3. show some kind of dirty indicator when the table is out of date compared to the state of the uma.